Support compilerfolder and online environments

+++ AppHandling/Run-TestsInNavContainer.ps1
@@ -87,6 +87,7 @@
 function Run-TestsInBcContainer {
     Param (
         [string] $containerName = $bcContainerHelperConfig.defaultContainerName,
+        [string] $compilerFolder = '',
         [Parameter(Mandatory=$false)]
         [string] $tenant = "default",
         [Parameter(Mandatory=$false)]
@@ -141,23 +142,49 @@ function Run-TestsInBcContainer {
 
 $telemetryScope = InitTelemetryScope -name $MyInvocation.InvocationName -parameterValues $PSBoundParameters -includeParameters @()
 try {
-    
-    $customConfig = Get-BcContainerServerConfiguration -ContainerName $containerName
-    $navversion = Get-BcContainerNavversion -containerOrImageName $containerName
-    $version = [System.Version]($navversion.split('-')[0])
+
+    if ($containerName) {
+        $customConfig = Get-BcContainerServerConfiguration -ContainerName $containerName
+        $navversion = Get-BcContainerNavversion -containerOrImageName $containerName
+        $version = [System.Version]($navversion.split('-')[0])
+        $PsTestToolFolder = Join-Path $bcContainerHelperConfig.hostHelperFolder "Extensions\$containerName\PsTestTool"
+
+    }
+    elseif ($compilerFolder) {
+        $customConfig = $null
+        $symbolsFolder = Join-Path $compilerFolder "symbols"
+        $baseAppInfo = Get-AppJsonFromAppFile -appFile (Get-ChildItem -Path $symbolsFolder -Filter 'Microsoft_Base Application_*.*.*.*.app').FullName
+        $version = [Version]$baseAppInfo.version
+        $PsTestToolFolder = Join-Path ([System.IO.Path]::GetTempPath()) "$([Guid]::NewGuid().ToString())"
+        New-Item $PsTestToolFolder -ItemType Directory | Out-Null
+        $testDlls = Join-Path $compilerFolder "dlls/Test Assemblies/*.dll"
+        Copy-Item $testDlls -Destination $PsTestToolFolder -Force
+        Copy-Item -Path (Join-Path $PSScriptRoot "PsTestFunctions.ps1") -Destination $PsTestToolFolder -Force
+        Copy-Item -Path (Join-Path $PSScriptRoot "ClientContext.ps1") -Destination $PsTestToolFolder -Force
+    }
+    else {
+        throw "You must specify either containerName or compilerFolder"
+    }
 
     if ($bcAuthContext -and $environment) {
-        $response = Invoke-RestMethod -Method Get -Uri "$($bcContainerHelperConfig.baseUrl.TrimEnd('/'))/$($bcAuthContext.tenantID)/$environment/deployment/url"
-        if($response.status -ne 'Ready') {
-            throw "environment not ready, status is $($response.status)"
+        if ($environment -like 'https://*') {
+            $useUrl = $environment
         }
-        $useUrl = $response.data.Split('?')[0]
-        $tenant = ($response.data.Split('?')[1]).Split('=')[1]
-
-        if ($testPage) {
-            throw "You cannot specify testPage when running tests in an Online tenant"
+        else {
+            $response = Invoke-RestMethod -Method Get -Uri "$($bcContainerHelperConfig.baseUrl.TrimEnd('/'))/$($bcAuthContext.tenantID)/$environment/deployment/url"
+            if($response.status -ne 'Ready') {
+                throw "environment not ready, status is $($response.status)"
+            }
+            $useUrl = $response.data
+            if ($testPage) {
+                throw "You cannot specify testPage when running tests in an Online tenant"
+            }
+            $testPage = 130455
         }
-        $testPage = 130455
+        $uri = [Uri]::new($useUrl)
+        $useUrl = $useUrl.Split('?')[0]
+        $dict = [System.Web.HttpUtility]::ParseQueryString($uri.Query)
+        if ($dict['tenant']) { $tenant = $dict['tenant'] }
     }
     else {
         $clientServicesCredentialType = $customConfig.ClientServicesCredentialType
@@ -226,7 +253,6 @@ try {
         $credential = New-Object pscredential -ArgumentList $bcAuthContext.upn, (ConvertTo-SecureString -String $accessToken -AsPlainText -Force)
     }
 
-    $PsTestToolFolder = Join-Path $bcContainerHelperConfig.hostHelperFolder "Extensions\$containerName\PsTestTool"
     $PsTestFunctionsPath = Join-Path $PsTestToolFolder "PsTestFunctions.ps1"
     $ClientContextPath = Join-Path $PsTestToolFolder "ClientContext.ps1"
     $fobfile = Join-Path $PsTestToolFolder "PSTestToolPage.fob"
@@ -273,24 +299,30 @@ try {
         try
         {
             if ($connectFromHost) {
+                if ($PSVersionTable.PSVersion.Major -lt 7) {
+                    throw "Using ConnectFromHost requires PowerShell 7"
+                }
                 $newtonSoftDllPath = Join-Path $PsTestToolFolder "NewtonSoft.json.dll"
                 $clientDllPath = Join-Path $PsTestToolFolder "Microsoft.Dynamics.Framework.UI.Client.dll"
-    
-                Invoke-ScriptInBcContainer -containerName $containerName { Param([string] $myNewtonSoftDllPath, [string] $myClientDllPath)
-                
-                    if (!(Test-Path $myNewtonSoftDllPath)) {
-                        $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Management\NewtonSoft.json.dll"
-                        if (!(Test-Path $newtonSoftDllPath)) {
-                            $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\NewtonSoft.json.dll"
-                        }
-                        $newtonSoftDllPath = (Get-Item $newtonSoftDllPath).FullName
-                        Copy-Item -Path $newtonSoftDllPath -Destination $myNewtonSoftDllPath
+                if ($containerName) {
+                    if (!((Test-Path $newtonSoftDllPath) -and (Test-Path $clientDllPath))) {
+                        Invoke-ScriptInBcContainer -containerName $containerName { Param([string] $myNewtonSoftDllPath, [string] $myClientDllPath)
+                    
+                            if (!(Test-Path $myNewtonSoftDllPath)) {
+                                $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\Management\NewtonSoft.json.dll"
+                                if (!(Test-Path $newtonSoftDllPath)) {
+                                    $newtonSoftDllPath = "C:\Program Files\Microsoft Dynamics NAV\*\Service\NewtonSoft.json.dll"
+                                }
+                                $newtonSoftDllPath = (Get-Item $newtonSoftDllPath).FullName
+                                Copy-Item -Path $newtonSoftDllPath -Destination $myNewtonSoftDllPath
+                            }
+                            $clientDllPath = "C:\Test Assemblies\Microsoft.Dynamics.Framework.UI.Client.dll"
+                            if (!(Test-Path $myClientDllPath)) {
+                                Copy-Item -Path $clientDllPath -Destination $myClientDllPath
+                            }
+                        } -argumentList $newtonSoftDllPath, $clientDllPath
                     }
-                    $clientDllPath = "C:\Test Assemblies\Microsoft.Dynamics.Framework.UI.Client.dll"
-                    if (!(Test-Path $myClientDllPath)) {
-                        Copy-Item -Path $clientDllPath -Destination $myClientDllPath
-                    }
-                } -argumentList $newtonSoftDllPath, $clientDllPath
+                }
     
                 if ($useUrl) {
                     $publicWebBaseUrl = $useUrl.TrimEnd('/')
